Optimalizacia importov

+++ src/main/java/org/example/sportevent/SportEvent.java
@@ -1,11 +1,12 @@
 package org.example.sportevent;
 
 import lombok.Data;
+
 import java.time.LocalDateTime;
 
 @Data
 public class SportEvent {
-    private long eventId;
+    private Integer eventId;
     private String eventName;
     private LocalDateTime startTime;
     private String sportType;
